added test for issue #172, then moved all code related to effect prediction to sub-module 'effects'
