+++ test/test_effect_collection.py
@@ -17,7 +17,7 @@ Test properties of EffectCollection
 """
 
 from nose.tools import eq_
-from varcode import IncompleteTranscript, Substitution
+from varcode.effects import IncompleteTranscript, Substitution
 from .data import tcga_ov_variants, ov_wustle_variants
 tcga_ov_effects = tcga_ov_variants.effects()
 ov_wustle_effects = ov_wustle_variants.effects()
