+++ test/test_mouse.py
@@ -2,7 +2,8 @@ from __future__ import absolute_import
 
 from nose.tools import eq_
 
-from varcode import load_vcf, load_vcf_fast, Variant, Substitution
+from varcode import load_vcf, load_vcf_fast, Variant
+from varcode.effects import Substitution
 from pyensembl import Genome, EnsemblRelease, MAX_ENSEMBL_RELEASE
 from .data import data_path
