+++ varcode/common.py
@@ -16,8 +16,6 @@ from __future__ import print_function, division, absolute_import
 from collections import defaultdict
 
 from functools import wraps
-from Bio.Seq import Seq
-from six import string_types, text_type
 
 def groupby_field(records, field_name):
     """
@@ -32,7 +30,6 @@ def groupby_field(records, field_name):
         if value is not None:
             groups[value].append(record)
     return dict(groups)
-
 
 def memoize(fn):
     """Simple memoization decorator for functions and methods,
@@ -51,10 +48,3 @@ def memoize(fn):
             return memoized_values[key]
 
     return wrapped_fn
-
-def bio_seq_to_str(seq):
-    if isinstance(seq, Seq):
-        return str(seq)
-    elif isinstance(seq, string_types):
-        return text_type(seq)
-    return seq
