+++ varcode/effects/effect_classes.py
@@ -362,6 +362,35 @@ class AlternateStartCodon(Silent):
         return "alternate-start-codon (%s>%s)" % (
             self.ref_codon, self.alt_codon)
 
+class AlternateStopCodon(Silent):
+    """
+    Change the stop codon (e.g. TAG->TGA) without other alterations
+    to the coding sequence. Currently this also includes downstream
+    modifications to the 5' UTR, eventually we need to move to a model
+    where a single variant can be described by a collection of overlapping
+    modifications.
+    """
+    def __init__(
+            self,
+            variant,
+            transcript,
+            aa_ref,
+            ref_codon,
+            alt_codon):
+        Silent.__init__(
+            self,
+            variant=variant,
+            transcript=transcript,
+            aa_pos=len(transcript.protein_sequence),
+            aa_ref="*")
+        self.ref_codon = bio_seq_to_str(ref_codon)
+        self.alt_codon = bio_seq_to_str(alt_codon)
+
+    @property
+    def short_description(self):
+        return "alternate-stop-codon (%s>%s)" % (
+            self.ref_codon, self.alt_codon)
+
 class NonsilentCodingMutation(CodingMutation):
     """
     All coding mutations other than silent codon substitutions
