+++ varcode/effects/effect_collection.py
