+++ varcode/effects/transcript_helpers.py
