+++ varcode/variant_collection.py
@@ -19,7 +19,7 @@ from collections import OrderedDict
 import pandas as pd
 from sercol import Collection
 
-from .effect_collection import EffectCollection
+from .effects import EffectCollection
 from .common import memoize
 from .variant import variant_ascending_position_sort_key
